+++ kernel/src/utils_kernel/funciones-thread-safe/cambio-de-estado/cambio-estado-proceso.h
@@ -31,7 +31,7 @@
     
     void pasar_de_exec_a_exit(t_pcb* pcb);
 
-    t_pcb* _sacar_pcb_de_cola(int pid, t_estado* estado);
+    t_pcb* _sacar_pcb_de_cola(uint8_t pid, t_estado* estado);
 
     void _enviar_a_finalizar_proceso(t_pcb* proceso_a_finalizar);
